+++ gslp/SupervisionGenerator.h
@@ -22,15 +22,15 @@ class SupervisionGenerator {
   PackingModel Model;
   unsigned EnumCap;
   unsigned SamplesPerBlock;
-  unsigned C;
-  unsigned W;
+  float C;
+  float W;
   // Number of iterations we run MCTS
   unsigned NumIters;
 
 public:
   SupervisionGenerator(PolicyArchiver &Archiver, FrontierEvaluator *Evaluator,
                        PackingModel Model, unsigned EnumCap,
-                       unsigned SamplesPerBlock, unsigned C, unsigned W,
+                       unsigned SamplesPerBlock, float C, float W,
                        unsigned NumIters = 1000)
       : Archiver(Archiver), Evaluator(Evaluator), Model(Model),
         EnumCap(EnumCap), SamplesPerBlock(SamplesPerBlock), C(C),
